Resolve static analysis issues
